+++ src/main/java/br/com/starti/domain/entity/Endereco.java
@@ -28,7 +28,7 @@ public class Endereco implements Serializable {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id_endereco")
-	private int idEndereco;
+	private Long idEndereco;
 	
 	@NotBlank
 	@Column
